Refactor shared services and configurations
- Moved Supabase configuration to shared package.
- Refactored API client to use shared services.
- Centralized Firebase authentication and FCM services in shared package.
- Introduced application-level configuration constants.
- Added environment configuration management.
- Updated dependencies for shared package.
