+++ apps/driver/lib/core/config.dart
@@ -1,7 +1,1 @@
-class AppConfig {
-  static const Duration apiTimeout = Duration(seconds: 10);
-  static const Duration geocodeTimeout = Duration(seconds: 6);
-  static const Duration routeTimeout = Duration(seconds: 8);
-  static const Duration profileUpdateTimeout = Duration(seconds: 10);
-  static const Duration quickActionTimeout = Duration(seconds: 5);
-}
+export 'package:truxify_shared/src/config/app_config.dart';
