+++ packages/truxify_shared/lib/src/config/app_config.dart
@@ -1,0 +1,9 @@
+/// Shared application-level configuration constants.
+class AppConfig {
+  static const Duration apiTimeout = Duration(seconds: 10);
+  static const Duration geocodeTimeout = Duration(seconds: 6);
+  static const Duration routeTimeout = Duration(seconds: 8);
+  static const Duration profileUpdateTimeout = Duration(seconds: 10);
+  static const Duration quickActionTimeout = Duration(seconds: 5);
+  static const Duration syncTimeout = Duration(seconds: 15);
+}
