+++ packages/truxify_shared/lib/truxify_shared.dart
@@ -1,9 +1,27 @@
+// ── Config ──────────────────────────────────────────────────────────
+export 'src/config/app_config.dart';
+export 'src/config/env.dart';
+export 'src/config/firebase_config.dart';
+export 'src/config/supabase_config.dart';
+
+// ── Services ────────────────────────────────────────────────────────
+export 'src/services/api_client.dart';
+export 'src/services/auth_service.dart';
+export 'src/services/fcm_service.dart';
+
+// ── Models ──────────────────────────────────────────────────────────
 export 'src/models/faq.dart';
 export 'src/models/notification_item.dart';
 export 'src/models/support_ticket.dart';
+
+// ── Repositories ────────────────────────────────────────────────────
 export 'src/repositories/faq_repository.dart';
 export 'src/repositories/notification_repository.dart';
 export 'src/repositories/support_repository.dart';
+
+// ── Screens ─────────────────────────────────────────────────────────
 export 'src/screens/help_center_screen.dart';
 export 'src/screens/notifications_screen.dart';
 
+// ── Widgets ─────────────────────────────────────────────────────────
+export 'shimmer_widget.dart';
